backend/requirements.txt
django 
djangorestframework 
django-cors-headerscl
djangorestframework-simplejwt